# specs/001-todo-web-app/contracts/openapi.yaml
openapi: "3.1.0"
info:
  title: Todo App Phase II — Task API
  version: "1.0.0"
  description: >
    REST API for the Phase II multi-user todo web application.
    All endpoints require a valid JWT issued by Better Auth.
    The {user_id} path parameter must match the authenticated user's identity
    extracted from the JWT (sub claim). Mismatches return 403.

servers:
  - url: http://localhost:8000
    description: Local development

security:
  - BearerAuth: []

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        JWT issued by Better Auth on successful sign-in.
        Verified by FastAPI middleware using BETTER_AUTH_SECRET.

  schemas:
    TaskResponse:
      type: object
      required: [id, user_id, title, is_complete, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440000"
        user_id:
          type: string
          format: uuid
          example: "6ba7b810-9dad-11d1-80b4-00c04fd430c8"
        title:
          type: string
          minLength: 1
          maxLength: 200
          example: "Buy groceries"
        description:
          type: string
          maxLength: 1000
          nullable: true
          example: "Milk, eggs, bread"
        is_complete:
          type: boolean
          example: false
        created_at:
          type: string
          format: date-time
          example: "2026-03-26T10:00:00Z"
        updated_at:
          type: string
          format: date-time
          example: "2026-03-26T10:05:00Z"

    TaskListResponse:
      type: object
      required: [tasks, total]
      properties:
        tasks:
          type: array
          items:
            $ref: "#/components/schemas/TaskResponse"
        total:
          type: integer
          description: Total count of tasks returned (supports future pagination)
          example: 5

    TaskCreateRequest:
      type: object
      required: [title]
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
          example: "Buy groceries"
        description:
          type: string
          maxLength: 1000
          nullable: true
          example: "Milk, eggs, bread"

    TaskUpdateRequest:
      type: object
      minProperties: 1
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
          example: "Buy organic groceries"
        description:
          type: string
          maxLength: 1000
          nullable: true
          example: "From the farmers market"

    ErrorResponse:
      type: object
      required: [detail]
      properties:
        detail:
          type: string
          example: "Not authorized"

    ValidationErrorResponse:
      type: object
      required: [detail]
      properties:
        detail:
          type: array
          items:
            type: object
            properties:
              loc:
                type: array
                items:
                  type: string
              msg:
                type: string
              type:
                type: string

paths:
  /api/{user_id}/tasks:
    parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
        description: >
          Must match the authenticated user's identity (JWT sub claim).
          Returns 403 if mismatch.

    get:
      operationId: listTasks
      summary: List all tasks for the authenticated user
      tags: [Tasks]
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [pending]
          description: >
            Filter by status. "pending" returns only incomplete tasks.
            Omit to return all tasks.
      responses:
        "200":
          description: Successful — returns task list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskListResponse"
        "401":
          description: Missing or invalid authentication credential
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Authenticated user does not match {user_id} in path
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          description: Invalid query parameter value
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationErrorResponse"

    post:
      operationId: createTask
      summary: Create a new task for the authenticated user
      tags: [Tasks]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskCreateRequest"
      responses:
        "201":
          description: Task created — returns the new task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskResponse"
        "401":
          description: Missing or invalid authentication credential
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Authenticated user does not match {user_id} in path
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          description: Validation error (empty title, title too long, description too long)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationErrorResponse"

  /api/{user_id}/tasks/{task_id}:
    parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
        description: Must match the authenticated user's identity (JWT sub claim).
      - name: task_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
        description: Unique identifier of the task.

    get:
      operationId: getTask
      summary: Get a single task by ID
      tags: [Tasks]
      responses:
        "200":
          description: Task found and owned by authenticated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskResponse"
        "401":
          description: Missing or invalid authentication credential
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Authenticated user does not match {user_id} in path
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Task not found (or belongs to a different user — same response to prevent leakage)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    put:
      operationId: updateTask
      summary: Update title and/or description of a task
      tags: [Tasks]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskUpdateRequest"
      responses:
        "200":
          description: Task updated — returns the updated task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskResponse"
        "401":
          description: Missing or invalid authentication credential
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Authenticated user does not match {user_id} in path
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Task not found or not owned by authenticated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          description: Validation error (empty title, title too long, description too long)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationErrorResponse"

    delete:
      operationId: deleteTask
      summary: Permanently delete a task
      tags: [Tasks]
      responses:
        "204":
          description: Task deleted — no content returned
        "401":
          description: Missing or invalid authentication credential
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Authenticated user does not match {user_id} in path
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Task not found or not owned by authenticated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /api/{user_id}/tasks/{task_id}/complete:
    parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
        description: Must match the authenticated user's identity (JWT sub claim).
      - name: task_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
        description: Unique identifier of the task.

    patch:
      operationId: toggleTaskComplete
      summary: Toggle the completion status of a task (no request body)
      tags: [Tasks]
      requestBody:
        required: false
      responses:
        "200":
          description: Status toggled — returns the updated task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskResponse"
        "401":
          description: Missing or invalid authentication credential
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Authenticated user does not match {user_id} in path
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Task not found or not owned by authenticated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"